+++ src/SettingsDialog.h
@@ -22,6 +22,9 @@ public:
     bool thumbnailsVisible() const;
     bool wheelZoomEnabled() const;
     int interpolationMode() const;
+    int sortKey() const;
+    bool sortDescending() const;
+    bool confirmDelete() const;
 
     void setLanguageIndex(int idx);
     void setThemeIndex(int idx);
@@ -29,6 +32,9 @@ public:
     void setThumbnailsVisible(bool v);
     void setWheelZoomEnabled(bool v);
     void setInterpolationMode(int mode);
+    void setSortKey(int key);
+    void setSortDescending(bool descending);
+    void setConfirmDelete(bool v);
 
     // Key sequences
     QKeySequence keyNext() const;
@@ -40,6 +46,7 @@ public:
     QKeySequence keyRotateLeft() const;
     QKeySequence keyRotateRight() const;
     QKeySequence keyFullscreen() const;
+    QKeySequence keyDelete() const;
 
     void setKeyNext(const QKeySequence &ks);
     void setKeyPrev(const QKeySequence &ks);
@@ -50,6 +57,7 @@ public:
     void setKeyRotateLeft(const QKeySequence &ks);
     void setKeyRotateRight(const QKeySequence &ks);
     void setKeyFullscreen(const QKeySequence &ks);
+    void setKeyDelete(const QKeySequence &ks);
 
 public slots:
     void accept() override;
@@ -66,6 +74,9 @@ private:
     QCheckBox *m_thumbCheck = nullptr;
     QCheckBox *m_wheelZoomCheck = nullptr;
     QComboBox *m_interpolationCombo = nullptr;
+    QComboBox *m_sortKeyCombo = nullptr;
+    QComboBox *m_sortOrderCombo = nullptr;
+    QCheckBox *m_confirmDeleteCheck = nullptr;
     QColor m_bgColor;
 
     // Keys tab
@@ -78,6 +89,7 @@ private:
     QKeySequenceEdit *m_keyRotateLeft = nullptr;
     QKeySequenceEdit *m_keyRotateRight = nullptr;
     QKeySequenceEdit *m_keyFullscreen = nullptr;
+    QKeySequenceEdit *m_keyDelete = nullptr;
 };
 
 #endif // SETTINGSDIALOG_H